Small changes following faster global search pr
